requirements.txt
streamlit~=1.47.1
qdrant-client~=1.15.1
sentence-transformers~=5.0.0
scikit-learn~=1.7.1
plotly~=6.2.0
pandas~=2.3.1
numpy~=2.3.2
protobuf~=5.29.5
langchain~=0.3.27
llama-index-core==0.13.0
google-generativeai==0.8.5
flake8~=7.3.0
